# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Booking API",
        "description": "API documentation for Event Booking System",
        "version": "1.0.0"
    },
    "paths": {
        "/api/attendee/register": {
            "post": {
                "tags": [
                    "Attendee"
                ],
                "summary": "Register a new attendee",
                "operationId": "7a16c885267e89ec8e912d65284db168",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attendee registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You're registered successfully"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jane Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jane@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/attendee/events": {
            "post": {
                "tags": [
                    "Attendee"
                ],
                "summary": "List events available for attendee with filters",
                "operationId": "0fe9b24b6d55458856ad8a7c561555f3",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Search by event title",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "laravel"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Filter by country (in, uk, usa)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "in",
                                "uk",
                                "usa"
                            ],
                            "example": "in"
                        }
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "Start date filter (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-04-12"
                        }
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "End date filter (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-04-15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Events fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event list fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Laravel Conference"
                                                    },
                                                    "country": {
                                                        "type": "string",
                                                        "example": "in"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-12 09:00:00"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-12 17:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}